Refactored navigaton bar and url architecture
